backend-facenet/requirements.txt
aiofiles==0.7.0
aiohttp==3.7.4.post0
aiostream==0.4.3
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
cryptography==35.0.0
docker==5.0.3
facenet-pytorch==2.5.2
fastapi==0.68.2
filelock==3.3.1
google-auth==2.3.0
grpcio==1.41.0
h11==0.12.0
httptools==0.2.0
idna==3.3
jina==2.1.10
keras==2.6.0
kubernetes==18.20.0
lz4==3.1.1
mtcnn==0.1.1
multidict==5.2.0
numpy==1.21.3
oauthlib==3.1.1
opencv-python==4.5.4.58
pathspec==0.9.0
Pillow==8.4.0
pretty-errors==1.2.24
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.10.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-multipart==0.0.5
PyYAML==6.0
pyzmq==22.3.0
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.12.0
rsa==4.7.2
six==1.16.0
starlette==0.14.2
torch==1.10.0
torchvision==0.11.1
tornado==6.1
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websocket-client==1.2.1
websockets==10.0
yarl==1.7.0
